refactor: adjust packages
