+++ application/src/main/java/br/com/jpcchaves/application/category/DeleteCategoryGateway.java
@@ -1,4 +1,4 @@
-package br.com.jpcchaves.application.gateway.category;
+package br.com.jpcchaves.application.category;
 
 public interface DeleteCategoryGateway {
   void delete(Long id);
