+++ application/src/main/java/br/com/jpcchaves/application/category/UpdateCategoryGateway.java
@@ -1,4 +1,4 @@
-package br.com.jpcchaves.application.gateway.category;
+package br.com.jpcchaves.application.category;
 
 import br.com.jpcchaves.core.domain.Category;
