+++ application/src/main/java/br/com/jpcchaves/application/todo/CreateTodoUseCaseImpl.java
@@ -1,6 +1,6 @@
-package br.com.jpcchaves.application.usecaseimpl.todo;
+package br.com.jpcchaves.application.todo;
 
-import br.com.jpcchaves.application.gateway.todo.CreateTodoGateway;
+import br.com.jpcchaves.application.todo.CreateTodoGateway;
 import br.com.jpcchaves.core.domain.Todo;
 import br.com.jpcchaves.usecase.todo.CreateTodoUseCase;
