+++ application/src/main/java/br/com/jpcchaves/application/todo/DeleteTodoGateway.java
@@ -1,4 +1,4 @@
-package br.com.jpcchaves.application.gateway.todo;
+package br.com.jpcchaves.application.todo;
 
 public interface DeleteTodoGateway {
   void delete(Long id);
