+++ application/src/main/java/br/com/jpcchaves/application/todo/DeleteTodoUseCaseImpl.java
@@ -1,6 +1,6 @@
-package br.com.jpcchaves.application.usecaseimpl.todo;
+package br.com.jpcchaves.application.todo;
 
-import br.com.jpcchaves.application.gateway.todo.DeleteTodoGateway;
+import br.com.jpcchaves.application.todo.DeleteTodoGateway;
 import br.com.jpcchaves.usecase.todo.DeleteTodoUseCase;
 
 public class DeleteTodoUseCaseImpl implements DeleteTodoUseCase {
