+++ application/src/main/java/br/com/jpcchaves/application/todo/ListTodoPaginatedGateway.java
@@ -1,4 +1,4 @@
-package br.com.jpcchaves.application.gateway.todo;
+package br.com.jpcchaves.application.todo;
 
 import br.com.jpcchaves.core.domain.PaginatedResponse;
 import br.com.jpcchaves.core.domain.PaginationRequest;
