+++ infrastructure/src/main/java/br/com/jpcchaves/infrastructure/gatewayimpl/category/ListCategoriesGatewayImpl.java
@@ -1,6 +1,6 @@
 package br.com.jpcchaves.infrastructure.gatewayimpl.category;
 
-import br.com.jpcchaves.application.gateway.category.ListCategoriesGateway;
+import br.com.jpcchaves.application.category.ListCategoriesGateway;
 import br.com.jpcchaves.core.domain.Category;
 import br.com.jpcchaves.infrastructure.mapper.CategoryMapper;
 import br.com.jpcchaves.infrastructure.persistence.entity.CategoryEntity;
